requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.6.0=py311hca03da5_0
anyio=4.2.0=py311hca03da5_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.7.2=hca03da5_0
cachetools=5.3.3=pypi_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py311hca03da5_0
fastapi=0.103.0=py311hca03da5_0
google-ai-generativelanguage=0.6.6=pypi_0
google-api-core=2.19.1=pypi_0
google-api-python-client=2.137.0=pypi_0
google-auth=2.32.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.7.2=pypi_0
googleapis-common-protos=1.63.2=pypi_0
grpcio=1.64.1=pypi_0
grpcio-status=1.62.2=pypi_0
h11=0.14.0=py311hca03da5_0
httplib2=0.22.0=pypi_0
idna=3.7=py311hca03da5_0
libffi=3.4.4=hca03da5_1
ncurses=6.4=h313beb8_0
openssl=3.0.14=h80987f9_0
pip=24.0=py311hca03da5_0
proto-plus=1.24.0=pypi_0
protobuf=4.25.3=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pydantic=2.5.3=py311hca03da5_0
pydantic-core=2.14.6=py311hf0e4da2_0
pyparsing=3.1.2=pypi_0
python=3.11.9=hb885b13_0
python-dotenv=1.0.1=pypi_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
rsa=4.9=pypi_0
setuptools=69.5.1=py311hca03da5_0
sniffio=1.3.0=py311hca03da5_0
sqlite=3.45.3=h80987f9_0
starlette=0.27.0=py311hca03da5_0
tk=8.6.14=h6ba3021_0
tqdm=4.66.4=pypi_0
typing-extensions=4.11.0=py311hca03da5_0
typing_extensions=4.11.0=py311hca03da5_0
tzdata=2024a=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.20.0=py311hca03da5_0
wheel=0.43.0=py311hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
